feat: 필요한 dto 추가
