+++ src/main/java/com/example/aitalk/domain/level/LevelAssessmentService.java
@@ -1,6 +1,8 @@
 package com.example.aitalk.domain.level;
 
+import com.example.aitalk.domain.level.dto.AiDataRequestDTO;
+
 public class LevelAssessmentService {
-	public void saveAiAssessment(LevelAssessmentRequestDTO aiData) {
+	public void saveAiAssessment(AiDataRequestDTO aiData) {
 	}
 }
