+++ src/main/java/com/example/aitalk/domain/level/LevelController.java
@@ -1,7 +1,5 @@
 package com.example.aitalk.domain.level;
 
-import java.util.Map;
-
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -9,6 +7,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.example.aitalk.api.dto.CommonResponse;
+import com.example.aitalk.domain.level.dto.AiDataRequestDTO;
 import com.example.aitalk.global.util.ResponseUtil;
 
 import io.swagger.v3.oas.annotations.Operation;
@@ -26,7 +25,7 @@ public class LevelController {
 	@Operation(summary = "AI 학습 평가 데이터 수신 및 저장", description = "Flask 서버로부터 사용자 학습 평가 데이터를 받아 DB에 저장")
 	@ApiResponse(responseCode = "200", description = "성공")
 	@ApiResponse(responseCode = "400", description = "요청 데이터 오류 또는 사용자 없음")
-	public ResponseEntity<CommonResponse<Void>> receiveAiData(@Valid @RequestBody LevelAssessmentRequestDTO aiData) {
+	public ResponseEntity<CommonResponse<Void>> receiveAiData(@Valid @RequestBody AiDataRequestDTO aiData) {
 
 		levelAssessmentService.saveAiAssessment(aiData);
