+++ src/main/java/com/example/aitalk/domain/level/dto/UserProfileDTO.java
@@ -1,0 +1,11 @@
+package com.example.aitalk.domain.level.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class UserProfileDTO {
+	private String email;
+	private String name;
+}
